refactor: move table config into table module
