+++ rust/src/lib.rs
@@ -84,7 +84,6 @@ compile_error!(
 
 pub mod action;
 pub mod data_catalog;
-pub mod delta_config;
 pub mod errors;
 pub mod operations;
 pub mod schema;
@@ -97,8 +96,8 @@ pub mod delta_datafusion;
 #[cfg(all(feature = "arrow", feature = "parquet"))]
 pub mod writer;
 
+pub use self::config::*;
 pub use self::data_catalog::{get_data_catalog, DataCatalog, DataCatalogError};
-pub use self::delta_config::*;
 pub use self::schema::partitions::*;
 pub use self::schema::*;
 pub use self::table::*;
